news-alert-app/requirements.txt
requests==2.31.0
python-dateutil==2.9.0
rich==13.7.1
